docshell/test/mochitest/test_bug1740516.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test pageshow event order for iframe</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    SimpleTest.waitForExplicitFinish();

    function waitForPageShow(outer, inner) {
      return new Promise((resolve) => {
        let results = [];
        outer.addEventListener("message", ({ data: persisted }) => {
          results.push({ name: outer.name, persisted });
          if (results.length == 2) {
            resolve(results);
          }
        }, { once: true });
        inner.addEventListener("message", ({ data: persisted }) => {
          results.push({ name: inner.name, persisted });
          if (results.length == 2) {
            resolve(results);
          }
        }, { once: true });
      });
    }
    async function runTest() {
      let outerBC = new BroadcastChannel("bug1740516_1");
      let innerBC = new BroadcastChannel("bug1740516_1_inner");

      let check = waitForPageShow(outerBC, innerBC).then(([first, second]) => {
        is(first.name, "bug1740516_1_inner", "Should get pageShow from inner iframe page first.");
        ok(!first.persisted, "First navigation shouldn't come from BFCache.");
        is(second.name, "bug1740516_1", "Should get pageShow from outer page second.");
        ok(!second.persisted, "First navigation shouldn't come from BFCache.");
      }, () => {
        ok(false, "The promises should not be rejected.");
      });
      window.open("file_bug1740516_1.html", "", "noopener");
      await check;

      check = waitForPageShow(outerBC, innerBC).then(([first, second]) => {
        is(first.name, "bug1740516_1_inner", "Should get pageShow from inner iframe page first.");
        ok(first.persisted, "Second navigation should come from BFCache");
        is(second.name, "bug1740516_1", "Should get pageShow from outer page second.");
        ok(second.persisted, "Second navigation should come from BFCache");
      }, () => {
        ok(false, "The promises should not be rejected.");
      });
      outerBC.postMessage("navigate");
      await check;

      check = waitForPageShow(outerBC, innerBC).then(([first, second]) => {
        is(first.name, "bug1740516_1_inner", "Should get pageShow from inner iframe page first.");
        ok(!first.persisted, "Third navigation should not come from BFCache");
        is(second.name, "bug1740516_1", "Should get pageShow from outer page second.");
        ok(!second.persisted, "Third navigation should not come from BFCache");
      }, () => {
        ok(false, "The promises should not be rejected.");
      });
      outerBC.postMessage("block_bfcache_and_navigate");
      await check;

      outerBC.postMessage("close");

      outerBC.close();
      innerBC.close();

      SimpleTest.finish();
    }
  </script>
</head>
<body onload="runTest();">
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
